refactor: restructure vendor task components, add task overview and comment features, and implement task attachment handling
