+++ apps/app/src/app/[locale]/(app)/(dashboard)/[orgId]/vendors/[vendorId]/components/update-title-and-description-form.tsx
@@ -23,7 +23,7 @@ import type { z } from "zod";
 import { updateVendorAction } from "../actions/update-vendor-action";
 import { updateVendorSchema } from "../actions/schema";
 
-export function UpdateVendorForm({
+export function UpdateTitleAndDescriptionForm({
   vendor,
 }: {
   vendor: Vendor;
