+++ apps/app/src/app/[locale]/(app)/(dashboard)/[orgId]/vendors/[vendorId]/tasks/[taskId]/page.tsx
@@ -1,0 +1,7 @@
+export default function TaskPage() {
+    return (
+        <div>
+            Enter
+        </div>
+    );
+}
